parameter should be recursive, idea is there
